Performance issues [WIP] (#4)
* :art:

Add the parallel flag to alway notify the solvers the parallelization
scheme.

* :sparkles:

Add a basic interface for the mcmove! function which will be implemented
by most ensembles.

* :art: :recycle:

There were lots of type instabilities, which are now fixed and have
greatly reduce the allocation space.

* :construction:

Some minor adjustments, there are still performance issues when doing
the full simulation.

* :boom:

This is not working correctly, still have type instability problems. I
think this comes from the fact that the Box type is not typed
concretely.

* :bug:

It seems like the type instabilities that I was getting were not that
important, and the System type is actually concrete, so the problem is
within the simulation itself.

* :art:

Move files around to make it possible for the Simulation type to be
concrete.

* :sparkles: :construction:

It seems that most of the allocations came from the fact that the
functions were not being inferred correctly, so I decided to remove the
interfaces and instead the potentials already come with the
corresponding functions.

* :sparkles:

Final touches to this PR. The code seems to be working and the huge
allocations were removed. It is somewhat slow, but that can be taken
care of later.

* :white_check_mark:

Updated the tests for simulations. Still some other things to test,
depending on the Codecov.

* :fire:

There is no longer a need for lattice setups when we have a packing
implementation.

* :bookmark:

New minor version.

* :pencil2:

Removed the random_init flag as it is no longer supported.

* Bump patch version.