More performance (#17)
* :recycle:

A complete refactoring to allow for different kind of pairwise
computation of the particles.

* :art: :recycle:

Another refactoring. Lots of barrier functions, and some more code
streamlining. The N squared implementation should already be working. It
is time to test it.

* :construction:

Performance is still a problem. I need to check the N squared one, but
at least the cell implementation is very slow.

* :construction:

Last changes, it seems to be going somewhat faster, it is still
unexplicably slow. It is at the same level of the N squared method, but
this one is parallelized.

* :construction:

Still trying to make performance a little bit better. I have tried to do
some more function barriers, but to no avail.

* :construction_worker:

Don't need to use CompatHelper or TagBot for now.

* :coffin:

The examples are useless now that there is a different API. In the
future, examples will be in the documentation.

* :sparkles:

Added a new type for passing auxiliary stuff around, as well as the cell
list.

* :recycle: :coffin: :white_check_mark:

Updated code, refactored some other parts of the code, and also removed
unnecessary interfaces for the potential definitions. Updated the tests
to resemble this.

* :white_check_mark:

Added a simple test for the n-square pairwise computation.

* :sparkles: :poop:

Added a new field to the EnsembleOptions type similar to a cache to
avoid allocations when moving particles. This is not the best solution,
but works okay until a complete refactoring of the type system is in
place.

* :coffin:

Removed unused function barrier for the computation of energy values.

* :bookmark:

Bump minor version.