Streamlining of the current type system (#18)
* :recycle:

Don't use abstract types for the different kind of potentials.

* :recycle: :art:

Streamlined the cache types by making the Vec3D type mutable. This
simplifies the typing, in order to avoid having unncessary cache types
for each ensemble.

* :art: :recycle: :memo:

Added buch of docstrings. Tried to fix some performance issues with the
pairwise computation. However, this is not a priority, as this feature
will soon be depracated.

* :bookmark:

Bump version.

* :white_check_mark:

Updated code for tests to pass.